add reigh data tool
