+++ LICENSE
@@ -3,7 +3,7 @@ Version 0.2 - March 24, 2026
 
 Copyright (c) 2025-2026 Peter O'Malley
 
-TL;DR (not legally binding — see full text below)
+Plain-language summary
 
 - Free for anyone to use internally, including commercial companies. Use it
   on personal projects, in research, or as a tool inside your business —
@@ -12,6 +12,8 @@ TL;DR (not legally binding — see full text below)
   source (primary business assets under an OSI-approved license or OSNL).
 - Paid tiers only apply if you redistribute the Program in a product/service
   AND your company is not open source — see Section 2(c) for the tiers.
+
+The terms below are the license.
 
 Preamble
