+++ reigh_data.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+from artagents.reigh_data import main
+
+if __name__ == "__main__":
+    raise SystemExit(main())
